4단계 - 자동차 경주 피드백 반영 | refactor : MVC 패턴으로 패키지 분리
