+++ src/main/java/racinggame/domain/Car.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public class Car {
     private final String name;
