+++ src/main/java/racinggame/domain/MovableMoveStrategy.java
@@ -1,9 +1,9 @@
-package racinggame;
+package racinggame.domain;
 
-public class NonMovableMoveStrategy implements MoveStrategy {
+public class MovableMoveStrategy implements MoveStrategy {
     private static final int THRESHOLD = 4;
     public boolean canMove() {
-        int randomNumber = 3;
+        int randomNumber = 4;
         return randomNumber >= THRESHOLD;
     }
 }
