+++ src/main/java/racinggame/domain/MoveStrategy.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 public interface MoveStrategy {
     boolean canMove();
