+++ src/main/java/racinggame/domain/RandomMoveStrategy.java
@@ -1,4 +1,4 @@
-package racinggame;
+package racinggame.domain;
 
 import java.util.Random;
