+++ src/main/java/racinggame/view/ResultView.java
@@ -1,4 +1,6 @@
-package racinggame;
+package racinggame.view;
+
+import racinggame.domain.Car;
 
 import java.util.List;
