+++ src/test/java/racinggame/CarTest.java
@@ -2,6 +2,9 @@ package racinggame;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racinggame.domain.Car;
+import racinggame.domain.MovableMoveStrategy;
+import racinggame.domain.NonMovableMoveStrategy;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
