+++ src/test/java/racinggame/InputValidatorTest.java
@@ -3,6 +3,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racinggame.view.InputValidator;
 
 import static org.junit.jupiter.api.Assertions.*;
